♻️  Move match rule collection into separate type/mode
